+++ deptag/learning/learn.py
@@ -556,7 +556,7 @@ def train_command(args: settings.Settings):
 
                 with torch.amp.autocast(
                         "cpu" if device == torch.device("cpu") else "cuda",
-                        enabled=True, dtype=torch.bfloat16
+                        enabled=True, dtype=torch.float16
                         ):
                     outputs = model(**batch)
